+++ sources/calendar-twig-1.x/src/src/Aeon/Twig/CalendarExtension.php
@@ -7,7 +7,13 @@ namespace Aeon\Twig;
 use Aeon\Calendar\Exception\InvalidArgumentException;
 use Aeon\Calendar\Gregorian\Calendar;
 use Aeon\Calendar\Gregorian\DateTime;
+use Aeon\Calendar\Gregorian\Day;
+use Aeon\Calendar\Gregorian\Interval;
+use Aeon\Calendar\Gregorian\Month;
+use Aeon\Calendar\Gregorian\Time;
 use Aeon\Calendar\Gregorian\TimeZone;
+use Aeon\Calendar\Gregorian\Year;
+use Aeon\Calendar\TimeUnit;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFilter;
 use Twig\TwigFunction;
@@ -16,27 +22,37 @@ final class CalendarExtension extends AbstractExtension
 {
     private Calendar $calendar;
 
-    private string $defaultFormat;
+    private string $defaultDateTimeFormat;
 
-    private ?string $defaultTimeZone;
+    private string $defaultDayFormat;
 
-    public function __construct(Calendar $calendar, string $defaultFormat = 'Y-m-d H:i:s', string $defaultTimeZone = null)
+    private string $defaultTimeFormat;
+
+    public function __construct(Calendar $calendar, string $defaultDateTimeFormat = 'Y-m-d H:i:s', string $defaultDayFormat = 'Y-m-d', string $defaultTimeFormat = 'H:i:s')
     {
-        if (\is_string($defaultTimeZone)) {
-            if (!TimeZone::isValid($defaultTimeZone)) {
-                throw new InvalidArgumentException($defaultTimeZone . ' is not valid timezone.');
-            }
-        }
-
         $this->calendar = $calendar;
-        $this->defaultFormat = $defaultFormat;
-        $this->defaultTimeZone = $defaultTimeZone;
+        $this->defaultDateTimeFormat = $defaultDateTimeFormat;
+        $this->defaultDayFormat = $defaultDayFormat;
+        $this->defaultTimeFormat = $defaultTimeFormat;
     }
 
     public function getFilters() : array
     {
         return [
-            new TwigFilter('aeon_date', [$this, 'aeon_date']),
+            new TwigFilter('aeon_datetime_format', [$this, 'aeon_datetime_format']),
+            new TwigFilter('aeon_day_format', [$this, 'aeon_day_format']),
+            new TwigFilter('aeon_time_format', [$this, 'aeon_time_format']),
+            new TwigFilter('aeon_in_seconds', [$this, 'aeon_in_seconds']),
+            new TwigFilter('aeon_in_seconds_precise', [$this, 'aeon_in_seconds_precise']),
+            new TwigFilter('aeon_interval', [$this, 'aeon_interval']),
+            new TwigFilter('aeon_second', [$this, 'aeon_second']),
+            new TwigFilter('aeon_seconds', [$this, 'aeon_second']),
+            new TwigFilter('aeon_minute', [$this, 'aeon_minute']),
+            new TwigFilter('aeon_minutes', [$this, 'aeon_minute']),
+            new TwigFilter('aeon_hour', [$this, 'aeon_hour']),
+            new TwigFilter('aeon_hours', [$this, 'aeon_hour']),
+            new TwigFilter('aeon_day', [$this, 'aeon_day']),
+            new TwigFilter('aeon_days', [$this, 'aeon_day']),
         ];
     }
 
@@ -44,24 +60,99 @@ final class CalendarExtension extends AbstractExtension
     {
         return [
             new TwigFunction('aeon_now', [$this, 'aeon_now']),
+            new TwigFunction('aeon_current_time', [$this, 'aeon_current_time']),
+            new TwigFunction('aeon_current_day', [$this, 'aeon_current_day']),
+            new TwigFunction('aeon_current_month', [$this, 'aeon_current_month']),
+            new TwigFunction('aeon_current_year', [$this, 'aeon_current_year']),
+            new TwigFunction('aeon_interval_closed', [$this, 'aeon_interval_closed']),
+            new TwigFunction('aeon_interval_open', [$this, 'aeon_interval_open']),
+            new TwigFunction('aeon_interval_right_open', [$this, 'aeon_interval_right_open']),
+            new TwigFunction('aeon_interval_left_open', [$this, 'aeon_interval_left_open']),
         ];
     }
 
-    public function aeon_date(DateTime $dateTime, string $format = null, string $timezone = null) : string
+    public function aeon_datetime_format(DateTime $dateTime, string $format = null, string $timezone = null) : string
     {
-        $tz = \is_string($timezone)
-            ? $timezone
-            : (\is_string($this->defaultTimeZone) ? $this->defaultTimeZone : null);
+        $tz = (\is_string($timezone) && TimeZone::isValid($timezone))
+            ? new TimeZone($timezone)
+            : null;
 
         $fmt = \is_string($format)
             ? $format
-            : $this->defaultFormat;
+            : $this->defaultDateTimeFormat;
 
-        if (\is_string($tz)) {
-            return $dateTime->toTimeZone(new TimeZone($tz))->format($fmt);
+        if ($tz instanceof TimeZone) {
+            return $dateTime->toTimeZone($tz)->format($fmt);
         }
 
         return $dateTime->format($fmt);
+    }
+
+    public function aeon_time_format(Time $time, string $format = null) : string
+    {
+        $fmt = \is_string($format)
+            ? $format
+            : $this->defaultTimeFormat;
+
+        return $time->format($fmt);
+    }
+
+    public function aeon_day_format(Day $day, string $format = null) : string
+    {
+        $fmt = \is_string($format)
+            ? $format
+            : $this->defaultDayFormat;
+
+        return $day->format($fmt);
+    }
+
+    public function aeon_interval(string $interval) : Interval
+    {
+        switch (\strtolower($interval)) {
+            case 'open':
+                return Interval::open();
+            case 'closed':
+                return Interval::closed();
+            case 'right_open':
+            case 'right open':
+                return Interval::rightOpen();
+            case 'left_open':
+            case 'left open':
+                return Interval::leftOpen();
+
+            default:
+                throw new InvalidArgumentException('Invalid interval type: ' . $interval);
+        }
+    }
+
+    public function aeon_in_seconds_precise(TimeUnit $timeUnit) : string
+    {
+        return $timeUnit->inSecondsPrecise();
+    }
+
+    public function aeon_in_seconds(TimeUnit $timeUnit) : int
+    {
+        return $timeUnit->inSeconds();
+    }
+
+    public function aeon_second(int $seconds) : TimeUnit
+    {
+        return TimeUnit::seconds($seconds);
+    }
+
+    public function aeon_minute(int $minutes) : TimeUnit
+    {
+        return TimeUnit::minutes($minutes);
+    }
+
+    public function aeon_hour(int $hours) : TimeUnit
+    {
+        return TimeUnit::hours($hours);
+    }
+
+    public function aeon_day(int $days) : TimeUnit
+    {
+        return TimeUnit::days($days);
     }
 
     public function aeon_now(string $timezone = null) : DateTime
@@ -71,5 +162,45 @@ final class CalendarExtension extends AbstractExtension
         }
 
         return $this->calendar->now();
+    }
+
+    public function aeon_current_time() : Time
+    {
+        return $this->calendar->now()->time();
+    }
+
+    public function aeon_current_day() : Day
+    {
+        return $this->calendar->currentDay();
+    }
+
+    public function aeon_current_month() : Month
+    {
+        return $this->calendar->currentMonth();
+    }
+
+    public function aeon_current_year() : Year
+    {
+        return $this->calendar->currentYear();
+    }
+
+    public function aeon_interval_left_open() : Interval
+    {
+        return Interval::leftOpen();
+    }
+
+    public function aeon_interval_right_open() : Interval
+    {
+        return Interval::rightOpen();
+    }
+
+    public function aeon_interval_closed() : Interval
+    {
+        return Interval::closed();
+    }
+
+    public function aeon_interval_open() : Interval
+    {
+        return Interval::open();
     }
 }