+++ sources/calendar-twig-1.x/src/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.4.0] - 2020-10-13
+- Added possibility to format time, removed redundant default timezone #25
+
+## [0.3.0] - 2020-10-13
+- Added more filters and functions 
+
 ## [0.2.0] - 2020-08-18
 - updated aeon-php/calendar dependency to >=0.4.0